[query] rename util to common object
