+++ fusequery/query/src/datasources/system/numbers_table.rs
@@ -21,7 +21,7 @@ use common_planners::Statistics;
 use common_streams::SendableDataBlockStream;
 
 use crate::datasources::system::NumbersStream;
-use crate::datasources::util::generate_parts;
+use crate::datasources::Common;
 use crate::datasources::ITable;
 use crate::datasources::ITableFunction;
 use crate::sessions::FuseQueryContextRef;
@@ -97,7 +97,7 @@ impl ITable for NumbersTable {
             db: "system".to_string(),
             table: self.name().to_string(),
             schema: self.schema.clone(),
-            partitions: generate_parts(0, ctx.get_max_threads()?, total),
+            partitions: Common::generate_parts(0, ctx.get_max_threads()?, total),
             statistics: statistics.clone(),
             description: format!(
                 "(Read from system.{} table, Read Rows:{}, Read Bytes:{})",
