+++ fusequery/query/src/sql/mod.rs
@@ -8,11 +8,12 @@ mod plan_parser_test;
 mod sql_parser_test;
 
 mod plan_parser;
+mod sql_common;
 mod sql_parser;
 mod sql_statement;
-mod sql_util;
 
 pub use plan_parser::PlanParser;
+pub use sql_common::SqlCommon;
 pub use sql_parser::DfParser;
 pub use sql_statement::DfCreateDatabase;
 pub use sql_statement::DfCreateTable;
@@ -20,4 +21,3 @@ pub use sql_statement::DfExplain;
 pub use sql_statement::DfShowSettings;
 pub use sql_statement::DfShowTables;
 pub use sql_statement::DfStatement;
-pub use sql_util::*;
